refactor: replace legacy failure handling with a new domain_failure structure
