+++ lib/core/presentation/base_viewmodel/base_cubit.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:fpdart/fpdart.dart';
 import '../../../features/auth/domain/failures/failure.dart';
+import '../../domain/failure/domain_failure.dart';
 import '../exception/ui_errors.dart';
 
 abstract class BaseCubit<STATE> extends Cubit<STATE> {
@@ -62,8 +63,11 @@ abstract class BaseCubit<STATE> extends Cubit<STATE> {
       ValidationFailure(:final message, :final fieldErrors) =>
         ValidationUiError(message: message, fieldErrors: fieldErrors),
 
-      UnknownFailure(:final message) => UnknownUiError(message),
       AuthFailure(:final message) => GenericUiError(message),
+
+      UnknownFailure(:final message) => UnknownUiError(message),
+
+      _ => const UnknownUiError("Unknown error occurred")
     };
   }
