+++ lib/features/auth/domain/use_cases/complete_onboarding_use_case.dart
@@ -1,6 +1,5 @@
 import 'package:fpdart/fpdart.dart';
-import 'package:hr_management/features/auth/domain/failures/failure.dart';
-
+import '../../../../core/domain/failure/domain_failure.dart';
 import '../repository/on_boarding_repository.dart';
 
 class CompleteOnboardingUseCase {
