+++ lib/features/auth/domain/use_cases/forgot_password_use_case.dart
@@ -1,6 +1,6 @@
 import 'package:fpdart/fpdart.dart';
+import '../../../../core/domain/failure/domain_failure.dart';
 import '../entity/auth_type.dart';
-import '../failures/failure.dart';
 import '../repository/auth_repository.dart';
 
 class ForgotPasswordUseCase {
