+++ lib/features/auth/domain/use_cases/load_identifier_use_case.dart
@@ -1,6 +1,6 @@
 import 'package:fpdart/fpdart.dart';
-import 'package:hr_management/features/auth/domain/failures/failure.dart';
 import 'package:hr_management/features/auth/domain/repository/auth_repository.dart';
+import '../../../../core/domain/failure/domain_failure.dart';
 
 class LoadIdentifierUseCase {
   final AuthRepository _repository;
