+++ lib/features/auth/domain/use_cases/login_use_case.dart
@@ -1,8 +1,8 @@
 import 'package:fpdart/fpdart.dart';
 import 'package:hr_management/features/auth/domain/entity/user.dart';
-import 'package:hr_management/features/auth/domain/failures/failure.dart';
 import 'package:hr_management/features/auth/domain/repository/auth_repository.dart';
 
+import '../../../../core/domain/failure/domain_failure.dart';
 import '../entity/auth_type.dart';
