+++ lib/features/auth/domain/use_cases/reset_password_use_case.dart
@@ -1,6 +1,6 @@
 import 'package:fpdart/fpdart.dart';
+import '../../../../core/domain/failure/domain_failure.dart';
 import '../entity/user.dart';
-import '../failures/failure.dart';
 import '../repository/auth_repository.dart';
 
 class ResetPasswordUseCase {
