+++ lib/features/main_navigation/presentation/widgets/bottom_nav_bar.dart
@@ -1,7 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:iconsax/iconsax.dart';
-import '../../../../core/presentation/design_system/theme/helper/extention_colors.dart';
 import '../../../../core/presentation/design_system/theme/helper/theme_extention.dart';
 
 class BottomNavBar extends StatelessWidget {
