+++ pubspec.yaml
@@ -31,7 +31,7 @@ dependencies:
   easy_localization: ^3.0.8
   cached_network_image: ^3.3.1
   iconsax: ^0.0.8
-  file_picker: ^8.0.0
+  file_picker: ^10.3.9
   flutter_pdfview: ^1.4.3
   flutter_svg: ^2.2.3
   country_picker: ^2.0.27
@@ -39,7 +39,7 @@ dependencies:
   flutter_dotenv: ^6.0.0
   flutter_secure_storage: ^10.0.0
   get_it: ^9.2.0
-  dio: ^5.9.0
+  dio: ^5.9.1
   equatable: ^2.0.7
   logger: ^2.6.2
   go_router: ^17.0.1
